pedido de stock arreglado y funcionanado

+++ back/src/db/migrations/20240213110647-create-estadistica.js
@@ -10,13 +10,13 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       costo_total: {
-        type: Sequelize.INTEGER
+        type: Sequelize.BIGINT(100)
       },
       recaudacion: {
-        type: Sequelize.INTEGER
+        type: Sequelize.BIGINT(100)
       },
       profit: {
-        type: Sequelize.INTEGER
+        type: Sequelize.BIGINT(100)
       },
       createdAt: {
         allowNull: false,
